modified:   sleep_func.hpp

+++ sleep_func.hpp
@@ -1,10 +1,10 @@
 #ifdef WIN32
-    void sleep(int secs){
+    void sleep(float secs){
         Sleep(secs);
     }
 #else
-    void sleep(int secs){
+    void sleep(float secs){
         int microsecs = secs * 1000000;
         usleep(microsecs);
     }
-#endif
+#endif
